Increase the allowable number of leases
To prevent the possible wrapping of of the number chunk data leases, the reference count has been changed from a byte to a short. This will allow Broadcast exchanges to support a larger number of concurrent read requests for the same chunk.

(cherry picked from commit b7f8bc677cd4b8cdf80369a45fea5b1dabd277a5)

+++ starburst-buffer-service/starburst-buffer-data-server/src/main/java/io/starburst/stargate/buffer/data/execution/Chunk.java
@@ -196,7 +196,7 @@ public class Chunk
         @GuardedBy("this")
         private boolean releaseRequested;
         @GuardedBy("this")
-        private byte referenceCount;
+        private short referenceCount;
 
         public ChunkData(
                 MemoryAllocator memoryAllocator,
